[maint] Rename trial_stim to stim
[maint] Rename trial_stim_view to stim_view
