+++ lib/src/stim/stim_view.dart
